Hasura Actions GET
Currently Hasura actions are GET requests but they contain body. This is casuing datadog to spasm for good reason. I will move them to GET requests but we should do that when migrating the endpoints out of legacy. This is a temp fix for now to make them post requests.