[DEV-1068] Refactor useImage hook to handle error responses as text
This change refactors the POST request in the useImage hook by removing the explicit setting of the "Content-Type" header to "text/plain;charset=UTF-8" for the response. Additionally, the commit clarifies the processing of error responses as text, aligning it with the removal of the header to potentially address concerns regarding how the response is interpreted, though it does not address a specific vulnerability mentioned to be on line 78.